clean up code. test=develop
