+++ python/paddle/fluid/tests/unittests/dygraph_to_static/bert_dygraph_model.py
@@ -16,54 +16,11 @@ from __future__ import absolute_import
 from __future__ import division
 from __future__ import print_function
 
-import six
-import json
-import numpy as np
-
-import paddle
 import paddle.fluid as fluid
-from paddle.fluid.dygraph import Embedding, LayerNorm, Linear, to_variable, Layer, guard
+from paddle.fluid.dygraph import Embedding, Linear, Layer
 from paddle.fluid.dygraph.jit import dygraph_to_static_func
-
-
-class PrePostProcessLayer(Layer):
-    def __init__(self, process_cmd, d_model, dropout_rate, name):
-        super(PrePostProcessLayer, self).__init__()
-        self.process_cmd = process_cmd
-        self.functors = []
-        self.exec_order = ""
-
-        for cmd in self.process_cmd:
-            if cmd == "a":  # add residual connection
-                self.functors.append(lambda x, y: x + y if y is not None else x)
-                self.exec_order += "a"
-            elif cmd == "n":  # add layer normalization
-                self.functors.append(
-                    self.add_sublayer(
-                        "layer_norm_%d" % len(
-                            self.sublayers(include_sublayers=False)),
-                        LayerNorm(
-                            normalized_shape=d_model,
-                            param_attr=fluid.ParamAttr(
-                                name=name + "_layer_norm_scale",
-                                initializer=fluid.initializer.Constant(1.)),
-                            bias_attr=fluid.ParamAttr(
-                                name=name + "_layer_norm_bias",
-                                initializer=fluid.initializer.Constant(0.)))))
-                self.exec_order += "n"
-            elif cmd == "d":  # add dropout
-                if dropout_rate:
-                    self.functors.append(lambda x: fluid.layers.dropout(
-                        x, dropout_prob=dropout_rate, is_test=False))
-                    self.exec_order += "d"
-
-    def forward(self, x, residual=None):
-        for i, cmd in enumerate(self.exec_order):
-            if cmd == "a":
-                x = self.functors[i](x, residual)
-            else:
-                x = self.functors[i](x)
-        return x
+from transformer_dygraph_model import PrePostProcessLayer
+from transformer_dygraph_model import MultiHeadAttention
 
 
 class PositionwiseFeedForwardLayer(Layer):
@@ -97,126 +54,9 @@ class PositionwiseFeedForwardLayer(Layer):
         hidden = self._i2h(x)
         if self._dropout_rate:
             hidden = fluid.layers.dropout(
-                hidden,
-                dropout_prob=self._dropout_rate,
-                upscale_in_train="upscale_in_train",
-                is_test=False)
+                hidden, dropout_prob=self._dropout_rate, is_test=False)
         out = self._h2o(hidden)
         return out
-
-
-class MultiHeadAttentionLayer(Layer):
-    def __init__(self,
-                 d_key,
-                 d_value,
-                 d_model,
-                 n_head=1,
-                 dropout_rate=0.,
-                 cache=None,
-                 gather_idx=None,
-                 static_kv=False,
-                 param_initializer=None,
-                 name=""):
-        super(MultiHeadAttentionLayer, self).__init__()
-        self._n_head = n_head
-        self._d_key = d_key
-        self._d_value = d_value
-        self._d_model = d_model
-        self._dropout_rate = dropout_rate
-
-        self._q_fc = Linear(
-            input_dim=d_model,
-            output_dim=d_key * n_head,
-            param_attr=fluid.ParamAttr(
-                name=name + '_query_fc.w_0', initializer=param_initializer),
-            bias_attr=name + '_query_fc.b_0')
-
-        self._k_fc = Linear(
-            input_dim=d_model,
-            output_dim=d_key * n_head,
-            param_attr=fluid.ParamAttr(
-                name=name + '_key_fc.w_0', initializer=param_initializer),
-            bias_attr=name + '_key_fc.b_0')
-
-        self._v_fc = Linear(
-            input_dim=d_model,
-            output_dim=d_value * n_head,
-            param_attr=fluid.ParamAttr(
-                name=name + '_value_fc.w_0', initializer=param_initializer),
-            bias_attr=name + '_value_fc.b_0')
-
-        self._proj_fc = Linear(
-            input_dim=d_value * n_head,
-            output_dim=d_model,
-            param_attr=fluid.ParamAttr(
-                name=name + '_output_fc.w_0', initializer=param_initializer),
-            bias_attr=name + '_output_fc.b_0')
-
-    def forward(self, queries, keys, values, attn_bias):
-        # compute q ,k ,v
-        keys = queries if keys is None else keys
-        values = keys if values is None else values
-
-        q = self._q_fc(queries)
-        k = self._k_fc(keys)
-        v = self._v_fc(values)
-
-        # split head
-
-        q_hidden_size = q.shape[-1]
-        reshaped_q = fluid.layers.reshape(
-            x=q,
-            shape=[0, 0, self._n_head, q_hidden_size // self._n_head],
-            inplace=False)
-        transpose_q = fluid.layers.transpose(x=reshaped_q, perm=[0, 2, 1, 3])
-
-        k_hidden_size = k.shape[-1]
-        reshaped_k = fluid.layers.reshape(
-            x=k,
-            shape=[0, 0, self._n_head, k_hidden_size // self._n_head],
-            inplace=False)
-        transpose_k = fluid.layers.transpose(x=reshaped_k, perm=[0, 2, 1, 3])
-
-        v_hidden_size = v.shape[-1]
-        reshaped_v = fluid.layers.reshape(
-            x=v,
-            shape=[0, 0, self._n_head, v_hidden_size // self._n_head],
-            inplace=False)
-        transpose_v = fluid.layers.transpose(x=reshaped_v, perm=[0, 2, 1, 3])
-
-        scaled_q = fluid.layers.scale(x=transpose_q, scale=self._d_key**-0.5)
-        # scale dot product attention
-        product = fluid.layers.matmul(
-            #x=transpose_q,
-            x=scaled_q,
-            y=transpose_k,
-            transpose_y=True)
-        #alpha=self._d_model**-0.5)
-        if attn_bias is not None:
-            product += attn_bias
-        weights = fluid.layers.softmax(product)
-        if self._dropout_rate:
-            weights_droped = fluid.layers.dropout(
-                weights,
-                dropout_prob=self._dropout_rate,
-                dropout_implementation="upscale_in_train",
-                is_test=False)
-            out = fluid.layers.matmul(weights_droped, transpose_v)
-        else:
-            out = fluid.layers.matmul(weights, transpose_v)
-
-        # combine heads
-        if len(out.shape) != 4:
-            raise ValueError("Input(x) should be a 4-D Tensor.")
-        trans_x = fluid.layers.transpose(out, perm=[0, 2, 1, 3])
-        final_out = fluid.layers.reshape(
-            x=trans_x,
-            shape=[0, 0, trans_x.shape[2] * trans_x.shape[3]],
-            inplace=False)
-
-        # fc to output
-        proj_out = self._proj_fc(final_out)
-        return proj_out
 
 
 class EncoderSubLayer(Layer):
@@ -240,36 +80,15 @@ class EncoderSubLayer(Layer):
         self._preprocess_cmd = preprocess_cmd
         self._postprocess_cmd = postprocess_cmd
         self._prepostprocess_dropout = prepostprocess_dropout
-
         self._preprocess_layer = PrePostProcessLayer(
-            self._preprocess_cmd,
-            d_model,
-            prepostprocess_dropout,
-            name=name + "_pre_att")
-
-        self._multihead_attention_layer = MultiHeadAttentionLayer(
-            d_key,
-            d_value,
-            d_model,
-            n_head,
-            attention_dropout,
-            None,
-            None,
-            False,
-            param_initializer,
-            name=name + "_multi_head_att")
-
+            self._preprocess_cmd, d_model, prepostprocess_dropout)
+        self._multihead_attention_layer = MultiHeadAttention(
+            d_key, d_value, d_model, n_head, attention_dropout,
+            param_initializer)
         self._postprocess_layer = PrePostProcessLayer(
-            self._postprocess_cmd,
-            d_model,
-            self._prepostprocess_dropout,
-            name=name + "_post_att")
+            self._postprocess_cmd, d_model, self._prepostprocess_dropout)
         self._preprocess_layer2 = PrePostProcessLayer(
-            self._preprocess_cmd,
-            d_model,
-            self._prepostprocess_dropout,
-            name=name + "_pre_ffn")
-
+            self._preprocess_cmd, d_model, self._prepostprocess_dropout)
         self._positionwise_feed_forward = PositionwiseFeedForwardLayer(
             hidden_act,
             d_inner_hid,
@@ -277,24 +96,16 @@ class EncoderSubLayer(Layer):
             relu_dropout,
             param_initializer,
             name=name + "_ffn")
-
         self._postprocess_layer2 = PrePostProcessLayer(
-            self._postprocess_cmd,
-            d_model,
-            self._prepostprocess_dropout,
-            name=name + "_post_ffn")
+            self._postprocess_cmd, d_model, self._prepostprocess_dropout)
 
     def forward(self, enc_input, attn_bias):
         pre_process_multihead = self._preprocess_layer(enc_input)
-
         attn_output = self._multihead_attention_layer(pre_process_multihead,
                                                       None, None, attn_bias)
         attn_output = self._postprocess_layer(attn_output, enc_input)
-
         pre_process2_output = self._preprocess_layer2(attn_output)
-
         ffd_output = self._positionwise_feed_forward(pre_process2_output)
-
         return self._postprocess_layer2(ffd_output, attn_output)
 
 
@@ -322,8 +133,7 @@ class EncoderLayer(Layer):
         self._n_layer = n_layer
         self._hidden_act = hidden_act
         self._preprocess_layer = PrePostProcessLayer(
-            self._preprocess_cmd, 3, self._prepostprocess_dropout,
-            "post_encoder")
+            self._preprocess_cmd, 3, self._prepostprocess_dropout)
 
         for i in range(n_layer):
             self._encoder_sublayers.append(
@@ -402,7 +212,7 @@ class BertModelLayer(Layer):
             act="tanh")
 
         self.pre_process_layer = PrePostProcessLayer(
-            "nd", self._emb_size, self._prepostprocess_dropout, "")
+            "nd", self._emb_size, self._prepostprocess_dropout)
 
         self._encoder = EncoderLayer(
             hidden_act=self._hidden_act,
@@ -420,9 +230,6 @@ class BertModelLayer(Layer):
             param_initializer=self._param_initializer)
 
     def forward(self, src_ids, position_ids, sentence_ids, input_mask):
-        """
-        forward
-        """
         src_emb = self._src_emb(src_ids)
         pos_emb = self._pos_emb(position_ids)
         sent_emb = self._sent_emb(sentence_ids)
@@ -454,10 +261,6 @@ class BertModelLayer(Layer):
 
 
 class PretrainModelLayer(Layer):
-    """
-    pretrain model
-    """
-
     def __init__(self,
                  config,
                  return_pooled_out=True,
@@ -481,7 +284,7 @@ class PretrainModelLayer(Layer):
             config=self.config, return_pooled_out=True, use_fp16=self.use_fp16)
 
         self.pre_process_layer = PrePostProcessLayer(
-            "n", self._emb_size, self._prepostprocess_dropout, "pre_encoder")
+            "n", self._emb_size, self._prepostprocess_dropout)
 
         self.pooled_fc = Linear(
             input_dim=self._emb_size,
@@ -521,9 +324,6 @@ class PretrainModelLayer(Layer):
     @dygraph_to_static_func
     def forward(self, src_ids, position_ids, sentence_ids, input_mask,
                 mask_label, mask_pos, labels):
-        """
-        forward
-        """
         mask_pos = fluid.layers.cast(x=mask_pos, dtype='int32')
 
         enc_output, next_sent_feat = self.bert_layer(src_ids, position_ids,
